Setup conftest.py and utils.py
Arrange shared fixtures into conftest.py.
Arrange shared functions into utils.py.
