+++ src/tests/utils.py
@@ -1,0 +1,12 @@
+import json
+
+
+def post_json(client, url, json_dict):
+    """Send dictionary json_dict as a json to the specified url """
+    return client.post(url, data=json.dumps(json_dict),
+                       content_type='application/json')
+
+
+def json_of_response(response):
+    """Decode json from response"""
+    return json.loads(response.data.decode('utf8'))
